iPad site width bugfix
Media query for tables was resetting the padding which caused cards to
push site out of the screen in iPad (and possible other tablets). Also
padding: 0 was setting all content on tablets to start from pixel 0
which made it hard to read. I removed the media query now, works fine
in iPad.