Touch icons and iOS 7 web app meta tags
Touch icons for normal and retina resolution with iOS 7 icon sizes for
iPhone and iPad, meta tags to enable iOS full screen mode, title and
translucent status bar.

This commit also depends on the build script adjustment by Sebastian
because the PNG touch images are referenced in the meta tags and need
to be suffixed with hash string.